refactor: update tests to assert class names instead of string representations and remove unused test cases
